# api-docs/swagger.yaml
components:
  headers:
    CacheControl:
      schema:
        const: no-store
        type: string
    ContentTypeOptions:
      schema:
        const: nosniff
        type: string
    FrameOptions:
      schema:
        const: DENY
        type: string
...
    RateLimitReset:
      description: Optional validated nonnegative safe-integer reset epoch.
      schema:
        maximum: 9007199254740991
        minimum: 0
        type: integer
    ReferrerPolicy:
      schema:
        const: strict-origin-when-cross-origin
        type: string
    RetryAfter:
      description: Nonnegative safe-integer delay in seconds.
      schema:
        maximum: 9007199254740991
        minimum: 0
        type: integer
    Vary:
      description: Response selection variance.
      schema:
        type: string
...
    XRequestID:
      description: Selected immutable request correlation value.
      schema:
        maxLength: 128
        minLength: 1
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
        type: string
  parameters:
...
    Owner:
      in: path
      name: owner
      required: true
      schema:
        maxLength: 39
        minLength: 1
        pattern: ^[A-Za-z0-9](?:[A-Za-z0-9_-]{0,37}[A-Za-z0-9])?$
        type: string
...
    XRequestID:
      description: A missing, invalid, repeated, or comma-combined candidate is replaced
        with a generated request ID.
      in: header
      name: X-Request-ID
      required: false
      schema:
        maxLength: 128
        minLength: 1
        pattern: ^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$
        type: string
  schemas:
...
    ErrorDetail:
      additionalProperties: false
      properties:
        detail:
          maxLength: 200
          minLength: 1
          type: string
        source:
          $ref: '#/components/schemas/ErrorSource'
      required:
      - detail
      type: object
    ErrorSource:
      oneOf:
      - additionalProperties: false
        properties:
          pointer:
            maxLength: 256
            minLength: 1
            type: string
        required:
        - pointer
        type: object
      - additionalProperties: false
        properties:
          parameter:
            maxLength: 256
            minLength: 1
            type: string
        required:
        - parameter
        type: object
      - additionalProperties: false
        properties:
          header:
            maxLength: 256
            minLength: 1
            type: string
        required:
        - header
        type: object
...
    GitHubOwner:
      additionalProperties: false
      properties:
        avatarUrl:
          format: uri
          pattern: ^https?://
          type: string
        bio:
          minLength: 1
          type:
          - string
          - "null"
        blog:
          minLength: 1
          type:
          - string
          - "null"
        company:
          minLength: 1
          type:
          - string
          - "null"
        createdAt:
          $ref: '#/components/schemas/Timestamp'
        followers:
          $ref: '#/components/schemas/SafeInteger'
        following:
          $ref: '#/components/schemas/SafeInteger'
        htmlUrl:
          format: uri
          pattern: ^https?://
          type: string
        id:
          $ref: '#/components/schemas/SafeInteger'
        location:
          minLength: 1
          type:
          - string
          - "null"
        login:
          minLength: 1
          type: string
        name:
          minLength: 1
          type:
          - string
          - "null"
        publicRepos:
          $ref: '#/components/schemas/SafeInteger'
        type:
          minLength: 1
          type: string
        updatedAt:
          $ref: '#/components/schemas/Timestamp'
      required:
      - id
      - login
      - type
      - name
      - avatarUrl
      - htmlUrl
      - company
      - blog
      - location
      - bio
      - publicRepos
      - followers
      - following
      - createdAt
      - updatedAt
      type: object
...
    Problem_github_not_found:
      additionalProperties: false
      properties:
        code:
          const: github_not_found
          type: string
        detail:
          const: GitHub resource not found
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 404
          type: integer
        title:
          const: Not Found
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_github_rate_limit:
      additionalProperties: false
      properties:
        code:
          const: github_rate_limit
          type: string
        detail:
          const: GitHub rate limit exceeded
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 429
          type: integer
        title:
          const: Too Many Requests
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_github_timeout:
      additionalProperties: false
      properties:
        code:
          const: github_timeout
          type: string
        detail:
          const: GitHub request timed out
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 504
          type: integer
        title:
          const: Gateway Timeout
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_github_upstream:
      additionalProperties: false
      properties:
        code:
          const: github_upstream
          type: string
        detail:
          const: GitHub upstream response is invalid or unavailable
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 502
          type: integer
        title:
          const: Bad Gateway
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_internal_error:
      additionalProperties: false
      properties:
        code:
          const: internal_error
          type: string
        detail:
          const: Internal server error
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 500
          type: integer
        title:
          const: Internal Server Error
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_invalid_request:
      additionalProperties: false
      properties:
        code:
          const: invalid_request
          type: string
        detail:
          const: Request is malformed
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 400
          type: integer
        title:
          const: Bad Request
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
    Problem_not_acceptable:
      additionalProperties: false
      properties:
        code:
          const: not_acceptable
          type: string
        detail:
          const: No acceptable response representation is available
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 406
          type: integer
        title:
          const: Not Acceptable
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
...
    Problem_validation_failed:
      additionalProperties: false
      properties:
        code:
          const: validation_failed
          type: string
        detail:
          const: Request validation failed
          type: string
        errors:
          items:
            $ref: '#/components/schemas/ErrorDetail'
          maxItems: 32
          minItems: 1
          type: array
        status:
          const: 422
          type: integer
        title:
          const: Unprocessable Content
          type: string
        type:
          const: about:blank
          type: string
      required:
      - title
      - status
      - detail
      - code
      type: object
...
    SafeInteger:
      maximum: 9007199254740991
      minimum: 0
      type: integer
    Timestamp:
      format: date-time
      pattern: ^[0-9]{4}-(?:0[1-9]|1[0-2])-(?:0[1-9]|[12][0-9]|3[01])T(?:[01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9]\.[0-9]{3}Z$
      type: string
  securitySchemes:
    BearerAuth:
      bearerFormat: Firebase ID token
      description: Firebase ID token verified with revocation checking.
      scheme: bearer
      type: http
info:
  description: Portable REST contract implemented with Echo 5.3.
  title: Echo Playground Portable API
  version: 1.0.0
...
openapi: 3.1.2
paths:
...
  /v1/github/owners/{owner}:
    get:
      description: The query string is closed; unknown, repeated, or malformed parameters
        are rejected.
      operationId: getGitHubOwner
      parameters:
      - $ref: '#/components/parameters/XRequestID'
      - $ref: '#/components/parameters/Owner'
      responses:
        "200":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/GitHubOwner'
            application/json:
              schema:
                $ref: '#/components/schemas/GitHubOwner'
          description: Successful response
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "400":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_invalid_request'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_invalid_request'
          description: Controlled invalid_request failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "404":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_github_not_found'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_github_not_found'
          description: Controlled github_not_found failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "406":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_not_acceptable'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_not_acceptable'
          description: Controlled not_acceptable failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "422":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_validation_failed'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_validation_failed'
          description: Controlled validation_failed failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "429":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_github_rate_limit'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_github_rate_limit'
          description: Controlled github_rate_limit failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-RateLimit-Reset:
              $ref: '#/components/headers/RateLimitReset'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "500":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_internal_error'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_internal_error'
          description: Controlled internal_error failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "502":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_github_upstream'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_github_upstream'
          description: Controlled github_upstream failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
        "504":
          content:
            application/cbor:
              schema:
                $ref: '#/components/schemas/Problem_github_timeout'
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem_github_timeout'
          description: Controlled github_timeout failure
          headers:
            Cache-Control:
              $ref: '#/components/headers/CacheControl'
            Referrer-Policy:
              $ref: '#/components/headers/ReferrerPolicy'
            Vary:
              $ref: '#/components/headers/Vary'
            X-Content-Type-Options:
              $ref: '#/components/headers/ContentTypeOptions'
            X-Frame-Options:
              $ref: '#/components/headers/FrameOptions'
            X-Request-ID:
              $ref: '#/components/headers/XRequestID'
      security: []
      summary: Get a public GitHub owner
...